moved ReadVectorizedResolver into api.model package
