+++ server/pxf-hive/src/main/java/org/greenplum/pxf/plugins/hive/HiveORCVectorizedResolver.java
@@ -36,7 +36,7 @@ import org.apache.hadoop.hive.serde2.objectinspector.StructObjectInspector;
 import org.apache.hadoop.io.Text;
 import org.greenplum.pxf.api.OneField;
 import org.greenplum.pxf.api.OneRow;
-import org.greenplum.pxf.api.ReadVectorizedResolver;
+import org.greenplum.pxf.api.model.ReadVectorizedResolver;
 import org.greenplum.pxf.api.error.UnsupportedTypeException;
 import org.greenplum.pxf.api.io.DataType;
