+++ server/pxf-service/src/main/java/org/greenplum/pxf/service/bridge/ReadVectorizedBridge.java
@@ -21,7 +21,7 @@ package org.greenplum.pxf.service.bridge;
 
 import org.greenplum.pxf.api.OneField;
 import org.greenplum.pxf.api.OneRow;
-import org.greenplum.pxf.api.ReadVectorizedResolver;
+import org.greenplum.pxf.api.model.ReadVectorizedResolver;
 import org.greenplum.pxf.api.io.Writable;
 import org.greenplum.pxf.api.model.RequestContext;
 import org.greenplum.pxf.service.utilities.BasePluginFactory;
